fix: Quote paths with spaces in "getting started"
Shell quoting can be very complicated and system- and shell-
specific, so don't try too hard to always correctly quote;
most commonly, the most special character in a filename
is a space, so handle this common case. If the user has something
weird going on then presumably they won't blindly copy and paste.