test: Fix the system tests
There are two aspects to these changes.

First, we now use a temp directory as the project's workspace when
testing.  Instead of operating directly on the git working tree, files
are copied into a temp directory (maybe excluding some) and start VSC
there.  This gives us more control over how files are updated, and
avoids the possibility that we'll mess up the developer's environment by
running a destructive git command in the wrong directory.

Second, we need to carefully coordinate when the index and scan services
run. We do this by initially excluding particular files when creating
the workspace, then copying them in afterwards. For example, we don't
copy in *.appmap.json files until we've waited for the indexer to start
(by watching for the pidfile it creates).