fix: Use glob to find appmap.yml
When running tests on Travis (and maybe, in general, when running the
extension on Linux), workspace.findFiles sometimes fails to find
existing files. This causes some of the integration tests to fail:
ProjectStateServiceInstance can't find appmap.yml, and so doesn't think
agent is installed.

These changes switch to using glob to find appmap.yml, which correctly
determines whether or not the file is available.