fix: Remove duplicate findings from appearing in the workspace
The previous strategy would cache findings by workspace but had no way
of knowing if a new finding was completely new or updated. As a result
if a appmap-findings.json file was updated, the updated findings would
be imported and cached alongside the old findings from the same finding.
To fix this, we now dynamically grab the findings from the workspace.