feat: Add dashboard and details pages for Runtime Analysis (#519)
* fix: small fixes

Remove duplicate listeners in installGuideView.js, remove extraneous 1 in installGuideWebview.ts, and convert the render arrow function in appmapView.js to a method.

* feat: add filtering methods to findings index

* feat: add findings webviews

* feat: support opening findings overview from finding details page

* feat: include rule information in resolved findings

* refactor: retrieve rule information from resolved finding instead of markdown file

* fix: remove unused imports

* refactor: rename findingsByHash function

* refactor: use destructuring to reduce verbosity

* test: add system tests

* chore: update scanner and components

* Add some fallback labels to findings list items

Sometimes a location is not available. Maybe the path no longer exists
in this project or the AppMap was shared from another source. Instead of
`undefined`, attempt to fall back to something descriptive.

* fix: fix broken system tests

Co-authored-by: Dustin Byrne <dustinsbyrne@gmail.com>