feat: Look for JavaScript dependencies in lockfiles
Sometimes supported web or test frameworks are not direct
dependencies of the project but instead are transitively pulled
by another one; this is especially common with express.js.

AppMap should still work correctly with most of these projects.
To make this use case easier, make sure to look in package-lock.json
and yaml.lock in the project support detection in addition to package.json.

Fixes #711