fix: Record AppMaps instructions state is consistent across views
We're now looking at the total amount of AppMaps present instead of a persisted flag. This was already true in the
"Record AppMaps" instructions page, but now it's also reflected in the sidebar state.

Project state now listens for AppMap file events via an AppMapCollection. This is because the collection was also
listening for file changes via FileChangeEmitter, so if the project state updated first, it would count N-1 AppMaps in
the collection. As a result the AppMapCollection interface was changed slightly to indicate the workspace folder where
the change took place when notifying listeners of an update.