fix: AppMaps are now removed from the list after deletion
Once the index folder containing a `metadata.json` is deleted, a deletion event for that `metadata.json` (or
`classMap.json`, etc.) would never fire due to removal of the index directory. This change introduces the IndexJanitor,
which listens for FileSystemEvents to determine when to remove the directory. This way, file system events are
guaranteed to be emitted before fully removing the index.

Note that this only affects deletion via the `appmap.deleteAllAppMaps` command or right-click deletion in the AppMap
tree list. If the user deletes an index directory by other means (e.g. `rm -rf tmp/appmap`), AppMaps will likely still remain.