fix: Don't send `appmap:create` on initialization
The AppMapWatcher should be indicating whether or not the create event was fired due to it starting up or not. This
behavior has been restored and is now tested.

The alternative option would be to enroll the AppMapWatcher before attaching listeners, but the AppMapCollection depends
on those initialization events to populate the collection of AppMaps. IMO it's not worth changing that behavior right
now.