Expose AnalysisManager as an extension export
This is so it can be examined in tests.

It may seem it's not necessary to expose it since `AnalysisManager`
is a singleton. However, since VSCode loads the extension in a
sandbox, the `AnalysisManager` instance visible to the test code is
in fact a separate instance than the one that's part of the extension.

The only way to interact with the ACTUAL `AnalysisManager` is through
`vscode.Extension.exports` which is in fact a proxy object which
transparently RPCs into the sandbox.

Because the test code examined a different `AnalysisManager` instance
than expected, the various assertions failed; however, since the
test code used `console.assert` instead of `assert.ok`, they just
printed `Assertion failed` without actually causing the test to fail.

(I'm not bothering to fix that problem since I'm going to refactor the
tests anyway.)