ci: Don't bother mocking os.homedir
Presumably this is to direct the appmap.jar to a known location.
But these tests aren't actually verifying anything about the JAR.
And, every other instance of VSCode that gets launched is going
to also try and update that JAR file in $HOME/.appmap - so we may
as well let these tests do it too.