fix: Don't hang when Copilot errors on streaming
This can happen eg. when the message is too long. In this case, Copilot
sometimes only throws when the response stream is read, not when it's
created. Previously errors there were uncaught, which led to a broken
connection. Now the errors are reported to the client correctly.

Also, use 422 in this case; it's likely non-recoverable so it doesn't
make sense for the client to keep retrying.