Rely on configuration events to restart Navie
This change removes a global five second debounce on changes to
`appMap.commandLineEnvironment`. However, this behavior is not ideal in
settings where the configuration is changed programatically along with
other configuration which also restarts the service.

A 100ms grace period is added to ensure multiple programmatic changes to
configuration can be submitted before the process is restarted without
triggering a longer debounce timeout.